@@ -3,30 +3,27 @@ package ca.cmetcalfe.locationshare;
 import android.os.Bundle;
 import android.preference.ListPreference;
 import android.preference.Preference;
+import android.preference.Preference.OnPreferenceChangeListener;
 import android.preference.PreferenceActivity;
 import android.preference.PreferenceManager;
-import android.support.v7.widget.Toolbar;
 import android.view.LayoutInflater;
-import android.view.View;
 import android.widget.LinearLayout;
+
+import androidx.appcompat.widget.Toolbar;
 
 public class SettingsActivity extends PreferenceActivity {
 
-	private static Preference.OnPreferenceChangeListener prefsListener =
-            new Preference.OnPreferenceChangeListener() {
-                @Override
-                public boolean onPreferenceChange(Preference pref, Object value) {
-                    String valueString = value.toString();
+    private static OnPreferenceChangeListener prefsListener = (pref, value) -> {
+        String valueString = value.toString();
 
-                    if (pref instanceof ListPreference) {
-                        ListPreference listPreference = (ListPreference) pref;
-                        int index = listPreference.findIndexOfValue(valueString);
+        if (pref instanceof ListPreference) {
+            ListPreference listPreference = (ListPreference) pref;
+            int index = listPreference.findIndexOfValue(valueString);
 
-                        pref.setSummary(index >= 0 ? listPreference.getEntries()[index] : null);
-                    }
-                    return true;
-                }
-            };
+            pref.setSummary(index >= 0 ? listPreference.getEntries()[index] : null);
+        }
+        return true;
+    };
 
     // ----------------------------------------------------
     // Android Lifecycle
@@ -47,25 +44,19 @@ public class SettingsActivity extends PreferenceActivity {
     private static void bindPreferenceSummaryToValue(Preference pref) {
         pref.setOnPreferenceChangeListener(prefsListener);
 
-        prefsListener.onPreferenceChange(pref,
-                                         PreferenceManager
-                                                 .getDefaultSharedPreferences(pref.getContext())
-                                                 .getString(pref.getKey(), ""));
+        prefsListener.onPreferenceChange(pref, PreferenceManager
+                .getDefaultSharedPreferences(pref.getContext())
+                .getString(pref.getKey(), ""));
     }
 
     // ----------------------------------------------------
     // Helper functions
     // ----------------------------------------------------
     private void setToolbar() {
-        LinearLayout root = (LinearLayout)findViewById(android.R.id.list).getParent().getParent().getParent();
+        LinearLayout root = (LinearLayout) findViewById(android.R.id.list).getParent().getParent().getParent();
         Toolbar toolbar = (Toolbar) LayoutInflater.from(this).inflate(R.layout.toolbar, root, false);
         root.addView(toolbar, 0);
         toolbar.setTitle(R.string.settings);
-        toolbar.setNavigationOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View v) {
-                finish();
-            }
-        });
+        toolbar.setNavigationOnClickListener(v -> finish());
     }
 }
