@@ -1,29 +1,22 @@
 package ca.cmetcalfe.locationshare;
 
 import android.Manifest;
-import android.app.AlertDialog;
 import android.content.ClipData;
 import android.content.ClipboardManager;
-import android.content.DialogInterface;
-import android.content.pm.PackageManager;
-import android.os.Build;
-import android.os.Bundle;
-
-import java.text.MessageFormat;
-import java.util.Locale;
-
 import android.content.Context;
+import android.content.DialogInterface;
+import android.content.DialogInterface.OnClickListener;
 import android.content.Intent;
+import android.content.pm.PackageManager;
 import android.location.Location;
 import android.location.LocationListener;
 import android.location.LocationManager;
 import android.net.Uri;
+import android.os.Build;
+import android.os.Bundle;
 import android.os.SystemClock;
 import android.preference.PreferenceManager;
 import android.provider.Settings;
-import android.support.annotation.NonNull;
-import android.support.v7.app.AppCompatActivity;
-import android.support.v7.widget.Toolbar;
 import android.util.Log;
 import android.view.Menu;
 import android.view.MenuItem;
@@ -32,6 +25,13 @@ import android.widget.Button;
 import android.widget.ProgressBar;
 import android.widget.TextView;
 import android.widget.Toast;
+
+import java.text.MessageFormat;
+import java.util.Locale;
+
+import androidx.annotation.NonNull;
+import androidx.appcompat.app.AlertDialog.Builder;
+import androidx.appcompat.app.AppCompatActivity;
 
 
 public class MainActivity extends AppCompatActivity {
@@ -55,12 +55,15 @@ public class MainActivity extends AppCompatActivity {
         public void onLocationChanged(Location loc) {
             updateLocation(loc);
         }
+
         public void onProviderEnabled(String provider) {
             updateLocation();
         }
+
         public void onProviderDisabled(String provider) {
             updateLocation();
         }
+
         public void onStatusChanged(String provider, int status, Bundle extras) {
         }
     };
@@ -72,7 +75,7 @@ public class MainActivity extends AppCompatActivity {
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-        setSupportActionBar((Toolbar)findViewById(R.id.toolbar));
+        setSupportActionBar(findViewById(R.id.toolbar));
         setTitle(R.string.app_name);
 
         // Display area
@@ -89,7 +92,7 @@ public class MainActivity extends AppCompatActivity {
         // Set default values for preferences
         PreferenceManager.setDefaultValues(this, R.xml.preferences, false);
 
-        locManager = (LocationManager)getSystemService(LOCATION_SERVICE);
+        locManager = (LocationManager) getSystemService(LOCATION_SERVICE);
     }
 
     @Override
@@ -160,14 +163,14 @@ public class MainActivity extends AppCompatActivity {
     // ----------------------------------------------------
     // DialogInterface Listeners
     // ----------------------------------------------------
-    private class onClickShareListener implements DialogInterface.OnClickListener {
+    private class onClickShareListener implements OnClickListener {
         @Override
         public void onClick(DialogInterface dialog, int i) {
             shareLocationText(formatLocation(lastLocation, getResources().getStringArray(R.array.link_options)[i]));
         }
     }
 
-    private class onClickCopyListener implements DialogInterface.OnClickListener {
+    private class onClickCopyListener implements OnClickListener {
         @Override
         public void onClick(DialogInterface dialog, int i) {
             copyLocationText(formatLocation(lastLocation, getResources().getStringArray(R.array.link_options)[i]));
@@ -181,7 +184,7 @@ public class MainActivity extends AppCompatActivity {
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         this.getMenuInflater().inflate(R.menu.menu_toolbar, menu);
-        return  true;
+        return true;
     }
 
     @Override
@@ -208,7 +211,7 @@ public class MainActivity extends AppCompatActivity {
         String linkChoice = PreferenceManager.getDefaultSharedPreferences(this).getString("prefLinkType", "");
 
         if (linkChoice.equals(getResources().getString(R.string.always_ask))) {
-            new AlertDialog.Builder(this).setTitle(R.string.choose_link)
+            new Builder(this).setTitle(R.string.choose_link)
                     .setCancelable(true)
                     .setItems(R.array.link_names, new onClickShareListener())
                     .create()
@@ -226,7 +229,7 @@ public class MainActivity extends AppCompatActivity {
         String linkChoice = PreferenceManager.getDefaultSharedPreferences(this).getString("prefLinkType", "");
 
         if (linkChoice.equals(getResources().getString(R.string.always_ask))) {
-            new AlertDialog.Builder(this).setTitle(R.string.choose_link)
+            new Builder(this).setTitle(R.string.choose_link)
                     .setCancelable(true)
                     .setItems(R.array.link_names, new onClickCopyListener())
                     .create()
@@ -256,7 +259,7 @@ public class MainActivity extends AppCompatActivity {
     // ----------------------------------------------------
     // Helper functions
     // ----------------------------------------------------
-    public void shareLocationText(String string){
+    public void shareLocationText(String string) {
         Intent intent = new Intent();
         intent.setAction(Intent.ACTION_SEND);
         intent.putExtra(Intent.EXTRA_TEXT, string);
@@ -264,14 +267,13 @@ public class MainActivity extends AppCompatActivity {
         startActivity(Intent.createChooser(intent, getString(R.string.share_location_via)));
     }
 
-    public void copyLocationText(String string){
-        ClipboardManager clipboard = (ClipboardManager)getSystemService(Context.CLIPBOARD_SERVICE);
-        if (clipboard != null){
+    public void copyLocationText(String string) {
+        ClipboardManager clipboard = (ClipboardManager) getSystemService(Context.CLIPBOARD_SERVICE);
+        if (clipboard != null) {
             ClipData clip = ClipData.newPlainText(getString(R.string.app_name), string);
             clipboard.setPrimaryClip(clip);
             Toast.makeText(getApplicationContext(), R.string.copied, Toast.LENGTH_SHORT).show();
-        }
-        else {
+        } else {
             Log.e(TAG, "Failed to get the clipboard service");
             Toast.makeText(getApplicationContext(), R.string.clipboard_error, Toast.LENGTH_SHORT).show();
         }
@@ -303,7 +305,6 @@ public class MainActivity extends AppCompatActivity {
         } else {
             return SystemClock.elapsedRealtime() - location.getElapsedRealtimeNanos() < 30e9;
         }
-
     }
 
     private String getAccuracy(Location location) {
